Makefile: Add possibility to uninstall spdk
This patch adds 'uninstall' target to makefile.
'make uninstall' will remove spdk_tgt app, headers, libraries
and shared libraries from system directories defined by $DESTDIR.
Additionaly, if there will be any empty directories left after
this operation, they will be removed as well.

+++ autobuild.sh
@@ -110,9 +110,19 @@ timing_enter make_install
 rm -rf /tmp/spdk
 mkdir /tmp/spdk
 $MAKE $MAKEFLAGS install DESTDIR=/tmp/spdk prefix=/usr
-ls -lR /tmp/spdk
-rm -rf /tmp/spdk
 timing_exit make_install
+
+# Test 'make uninstall'
+timing_enter make_uninstall
+$MAKE $MAKEFLAGS uninstall DESTDIR=/tmp/spdk prefix=/usr
+if [ "$(ls -A /tmp/spdk)" ];
+then
+	rm -rf /tmp/spdk
+	mkdir /tmp/spdk
+	echo "Make uninstall failed"
+	exit 1
+fi
+timing_exit make_uninstall
 
 timing_enter doxygen
 if [ $SPDK_BUILD_DOC -eq 1 ] && hash doxygen; then
